Refactor, and added time out state
